+++ packages/components/src/components/input-time-zone/input-time-zone.e2e.ts
@@ -1,5 +1,8 @@
 import { describe } from "vitest";
 import { labelable } from "../../tests/commonTests";
+import { ComponentTestTokens, themed } from "../../tests/commonTests/themed";
+import { html } from "../../../support/formatting";
+import { CSS as ComboboxCSS } from "../combobox/resources";
 import { mockConsole } from "../../tests/utils/logging";
 
 /*
@@ -13,4 +16,21 @@ mockConsole();
 
 describe("labelable", () => {
   labelable("calcite-input-time-zone");
+});
+
+describe("theme", () => {
+  const inputTimeZoneTokens: ComponentTestTokens = {
+    "--calcite-input-time-zone-corner-radius": [
+      {
+        shadowSelector: "calcite-combobox",
+        targetProp: "--calcite-combobox-corner-radius",
+      },
+      {
+        shadowSelector: `calcite-combobox >>> .${ComboboxCSS.wrapper}`,
+        targetProp: "borderRadius",
+      },
+    ],
+  };
+
+  themed(html`<calcite-input-time-zone></calcite-input-time-zone>`, inputTimeZoneTokens);
 });