+++ packages/components/src/components/table/table.e2e.ts
@@ -1,4 +1,4 @@
-import { newE2EPage } from "@arcgis/lumina-compiler/puppeteerTesting";
+import { E2EPage, newE2EPage } from "@arcgis/lumina-compiler/puppeteerTesting";
 import { describe, expect, it } from "vitest";
 import { html } from "../../../support/formatting";
 import { themed } from "../../tests/commonTests";
@@ -15,6 +15,23 @@ import type { TableHeader } from "../table-header/table-header";
 import type { TableCell } from "../table-cell/table-cell";
 import { TableRow } from "../table-row/table-row";
 import { SLOTS } from "./resources";
+
+async function getFocusedTableCellClassList(
+  page: E2EPage,
+  rowId: string,
+  cellSelector: "td" | "th",
+): Promise<string[]> {
+  return page.$eval(
+    `#${rowId}`,
+    (el, cellSelector) => {
+      const activeElement = el.shadowRoot!.activeElement as HTMLElement;
+      const cell = activeElement.shadowRoot!.querySelector<HTMLElement>(cellSelector)!;
+
+      return Array.from(cell.classList);
+    },
+    cellSelector,
+  );
+}
 
 describe("selection modes", () => {
   it("selection mode single allows one or no rows to be selected", async () => {
@@ -131,7 +148,6 @@ describe("selection modes", () => {
     expect(await element.getProperty("selectedItems")).toEqual([]);
     await selectedItemAsserter([]);
   });
-
   it("selection mode multiple allows one, multiple, or no rows to be selected", async () => {
     const page = await newE2EPage();
     await page.setContent(
@@ -1737,11 +1753,9 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("Tab");
     await page.waitForChanges();
 
-    expect(
-      await page.$eval(`#${rowHead.id}`, async (el) => {
-        return el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList;
-      }),
-    ).toEqual({ "0": HEADER_CSS.selectionCell, "1": HEADER_CSS.multipleSelectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell, HEADER_CSS.multipleSelectionCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
@@ -1753,27 +1767,21 @@ describe("keyboard navigation", () => {
 
     await page.keyboard.press("Home");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.selectionCell, "1": HEADER_CSS.multipleSelectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell, HEADER_CSS.multipleSelectionCell]),
+    );
 
     await page.keyboard.press("PageDown");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowFoot.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList,
-      ),
-    ).toEqual({ "0": CELL_CSS.footerCell, "1": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, rowFoot.id, "td")).toEqual(
+      expect.arrayContaining([CELL_CSS.footerCell, HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowUp");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row3.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, row3.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
@@ -1789,12 +1797,9 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("Home");
     await page.keyboard.up("ControlLeft");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.selectionCell, "1": HEADER_CSS.multipleSelectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell, HEADER_CSS.multipleSelectionCell]),
+    );
   });
 
   it("navigates correctly when number column present", async () => {
@@ -1831,12 +1836,9 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("Tab");
     await page.waitForChanges();
 
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
@@ -1848,27 +1850,21 @@ describe("keyboard navigation", () => {
 
     await page.keyboard.press("Home");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("PageDown");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowFoot.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList,
-      ),
-    ).toEqual({ "0": CELL_CSS.footerCell, "1": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowFoot.id, "td")).toEqual(
+      expect.arrayContaining([CELL_CSS.footerCell, HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("ArrowUp");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row3.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, row3.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
@@ -1884,12 +1880,9 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("Home");
     await page.keyboard.up("ControlLeft");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
   });
 
   it("navigates correctly when number and selection column present numbered", async () => {
@@ -1926,22 +1919,16 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("Tab");
     await page.waitForChanges();
 
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
 
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
@@ -1949,33 +1936,25 @@ describe("keyboard navigation", () => {
 
     await page.keyboard.press("ArrowLeft");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual([HEADER_CSS.selectionCell]);
 
     await page.keyboard.press("PageDown");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowFoot.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList,
-      ),
-    ).toEqual({ "0": CELL_CSS.footerCell, "1": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, rowFoot.id, "td")).toEqual(
+      expect.arrayContaining([CELL_CSS.footerCell, HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowUp");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row3.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, row3.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowLeft");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row3.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, row3.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
@@ -1987,12 +1966,9 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("Home");
     await page.keyboard.up("ControlRight");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
   });
 
   it("navigates correctly when pagination present and selection and number and first page displayed", async () => {
@@ -2033,22 +2009,16 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("Tab");
     await page.waitForChanges();
 
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
 
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.selectionCell, "1": HEADER_CSS.multipleSelectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell, HEADER_CSS.multipleSelectionCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
@@ -2056,27 +2026,21 @@ describe("keyboard navigation", () => {
 
     await page.keyboard.press("ArrowLeft");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.selectionCell, "1": HEADER_CSS.multipleSelectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell, HEADER_CSS.multipleSelectionCell]),
+    );
 
     await page.keyboard.press("PageDown");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowFoot.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList,
-      ),
-    ).toEqual({ "0": CELL_CSS.footerCell, "1": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, rowFoot.id, "td")).toEqual(
+      expect.arrayContaining([CELL_CSS.footerCell, HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowUp");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row2.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, row2.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
@@ -2090,23 +2054,17 @@ describe("keyboard navigation", () => {
 
     await page.keyboard.press("Home");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowFoot.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList,
-      ),
-    ).toEqual({ "0": CELL_CSS.footerCell, "1": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowFoot.id, "td")).toEqual(
+      expect.arrayContaining([CELL_CSS.footerCell, HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.down("ControlLeft");
     await page.keyboard.press("Home");
     await page.keyboard.up("ControlLeft");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
   });
 
   it("navigates correctly when pagination present, and selection and number and navigation to two other pages occurs", async () => {
@@ -2150,21 +2108,16 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("Tab");
     await page.waitForChanges();
 
-    expect(
-      await page.$eval(`#${rowHead.id}`, (el) => {
-        return el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList;
-      }),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
 
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
@@ -2172,45 +2125,39 @@ describe("keyboard navigation", () => {
 
     await page.keyboard.press("ArrowLeft");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowDown");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row1.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, row1.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowLeft");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row1.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, row1.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("PageDown");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowFoot.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList,
-      ),
-    ).toEqual({ "0": CELL_CSS.footerCell, "1": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowFoot.id, "td")).toEqual(
+      expect.arrayContaining([CELL_CSS.footerCell, HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("ArrowUp");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row2.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, row2.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row2.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, row2.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.down("ControlRight");
     await page.keyboard.press("End");
@@ -2220,23 +2167,17 @@ describe("keyboard navigation", () => {
 
     await page.keyboard.press("Home");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowFoot.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList,
-      ),
-    ).toEqual({ "0": CELL_CSS.footerCell, "1": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowFoot.id, "td")).toEqual(
+      expect.arrayContaining([CELL_CSS.footerCell, HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.down("ControlRight");
     await page.keyboard.press("Home");
     await page.keyboard.up("ControlRight");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.$eval(
       "calcite-table",
@@ -2259,45 +2200,39 @@ describe("keyboard navigation", () => {
 
     await page.keyboard.press("Home");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("ArrowDown");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row3.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, row3.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row3.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, row3.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("PageDown");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowFoot.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList,
-      ),
-    ).toEqual({ "0": CELL_CSS.footerCell, "1": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, rowFoot.id, "td")).toEqual(
+      expect.arrayContaining([CELL_CSS.footerCell, HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowUp");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row4.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, row4.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.selectionCell]),
+    );
 
     await page.keyboard.press("ArrowLeft");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row4.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, row4.id, "td")).toEqual(
+      expect.arrayContaining([HEADER_CSS.numberCell]),
+    );
 
     await page.keyboard.down("ControlRight");
     await page.keyboard.press("End");
@@ -2307,12 +2242,9 @@ describe("keyboard navigation", () => {
 
     await page.keyboard.press("Home");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowFoot.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList,
-      ),
-    ).toEqual({ "0": CELL_CSS.footerCell, "1": HEADER_CSS.numberCell });
+    expect(await getFocusedTableCellClassList(page, rowFoot.id, "td")).toEqual(
+      expect.arrayContaining([CELL_CSS.footerCell, HEADER_CSS.numberCell]),
+    );
   });
 
   it("navigates correctly when number and selection column present numbered and interaction-mode static - only focusing selection cells", async () => {
@@ -2350,58 +2282,44 @@ describe("keyboard navigation", () => {
     await page.keyboard.press("Tab");
     await page.waitForChanges();
 
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.selectionCell, "1": HEADER_CSS.multipleSelectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual([
+      HEADER_CSS.selectionCell,
+      HEADER_CSS.multipleSelectionCell,
+    ]);
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
 
     await page.keyboard.press("ArrowLeft");
     await page.waitForChanges();
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.selectionCell, "1": HEADER_CSS.multipleSelectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual([
+      HEADER_CSS.selectionCell,
+      HEADER_CSS.multipleSelectionCell,
+    ]);
 
     await page.keyboard.press("ArrowRight");
     await page.waitForChanges();
 
-    expect(
-      await page.$eval(
-        `#${rowHead.id}`,
-        (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("th")!.classList,
-      ),
-    ).toEqual({ "0": HEADER_CSS.selectionCell, "1": HEADER_CSS.multipleSelectionCell });
+    expect(await getFocusedTableCellClassList(page, rowHead.id, "th")).toEqual([
+      HEADER_CSS.selectionCell,
+      HEADER_CSS.multipleSelectionCell,
+    ]);
 
     await page.keyboard.press("Tab");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row1.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, row1.id, "td")).toEqual([HEADER_CSS.selectionCell]);
 
     await page.keyboard.press("Tab");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row2.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, row2.id, "td")).toEqual([HEADER_CSS.selectionCell]);
 
     await page.keyboard.press("Tab");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row3.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, row3.id, "td")).toEqual([HEADER_CSS.selectionCell]);
 
     await page.keyboard.press("ArrowUp");
     await page.waitForChanges();
-    expect(
-      await page.$eval(`#${row3.id}`, (el) => el.shadowRoot!.activeElement!.shadowRoot!.querySelector("td")!.classList),
-    ).toEqual({ "0": HEADER_CSS.selectionCell });
+    expect(await getFocusedTableCellClassList(page, row3.id, "td")).toEqual([HEADER_CSS.selectionCell]);
   });
 
   it("updates table-row's selected property correctly when calciteTabRowSelect event is emitted", async () => {
@@ -2424,8 +2342,8 @@ describe("keyboard navigation", () => {
     );
 
     async function selectRow(rowSelector: string): Promise<void> {
-      await page.$eval(rowSelector + " >>> calcite-table-cell:first-child", (el) => {
-        (el as TableCell["el"]).click();
+      await page.$eval(rowSelector + " >>> calcite-table-cell:first-child", (el: TableCell["el"]) => {
+        el.click();
       });
       await page.waitForChanges();
     }
@@ -2528,8 +2446,9 @@ describe("keyboard navigation", () => {
         </calcite-table>`,
         {
           "--calcite-table-border-color": {
-            shadowSelector: `.${TABLE_CSS.tableContainer}`,
-            targetProp: "borderColor",
+            selector: "#row-1",
+            shadowSelector: "tr",
+            targetProp: "--calcite-table-row-border-color",
           },
           "--calcite-table-corner-radius": {
             shadowSelector: `.${TABLE_CSS.tableContainer}`,
@@ -2829,8 +2748,8 @@ describe("keyboard navigation", () => {
           },
           "--calcite-table-row-border-color": {
             selector: "#row-1",
-            shadowSelector: "tr",
-            targetProp: "borderBlockEndColor",
+            shadowSelector: "calcite-table-cell >>> td",
+            targetProp: "borderBottomColor",
           },
         },
       );